update desc and bin

+++ rasgoudt/bin/publish-pypi.sh
@@ -1,3 +1,14 @@
+#!/bin/bash
+
+set -e
+
+if [ 1 != $# ]; then
+    echo "usage: $0 index"
+    echo "Index values: pypi pypitest"
+    exit 1;
+fi
+PYPI_INDEX="$1"
+
 # Remove old build artifacts
 rm -rf dist/*
 
@@ -5,4 +16,4 @@ rm -rf dist/*
 python setup.py sdist bdist_wheel
 
 # Upload artifacts to pypi
-python -m twine upload --verbose -r pypi dist/*
+python -m twine upload --verbose  -r "$PYPI_INDEX" dist/*
